Rename `ProjectValidator` to `ProjectMutator`
