+++ plugin/pkg/project/mutator/validator_suite_test.go
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package validator_test
+package mutator_test
 
 import (
 	"testing"
